Move HashingMetadataStore to its own file
